requirements.txt
absl-py==2.3.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
flatbuffers==25.2.10
fonttools==4.58.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.5.1
namex==0.1.0
numba==0.60.0
numpy==2.0.2
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.5
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
shap==0.47.2
six==1.17.0
slicer==0.0.8
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
